Separating variants into their own files.
