+++ packages/experiments/src/components/Button/ButtonVariants.types.ts
@@ -1,0 +1,7 @@
+import { IButtonProps } from './Button.types';
+
+export interface IButtonVariantProps extends IButtonProps {
+  text?: string;
+}
+
+export type ButtonVariantsType = (props: IButtonVariantProps) => JSX.Element;
