+++ packages/experiments/src/components/Button/PrimaryButton.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { Button } from './Button';
+import { ButtonVariantsType } from './ButtonVariants.types';
+
+export const PrimaryButton: ButtonVariantsType = props => {
+  const { text, ...rest } = props;
+
+  return <Button primary content={text} {...rest} />;
+};
