+++ packages/experiments/src/components/Button/index.ts
@@ -1,3 +1,5 @@
 export * from './Button';
 export * from './Button.types';
-export * from './ButtonVariants';
+export * from './ButtonVariants.types';
+export * from './DefaultButton';
+export * from './PrimaryButton';
